+++ docs/resources/changelog.md
@@ -13,7 +13,10 @@ _During the beta period, these restrictions may be relaxed in the event of a mis
 - Fixed a bug that prevented forms from submitting when pressing <kbd>Enter</kbd> inside of an `<sl-input>` [#700](https://github.com/shoelace-style/shoelace/issues/700)
 - Fixed a bug in `<sl-input>` that prevented the `valueAsDate` and `valueAsNumber` properties from working when set before the component was initialized
 - Improved `autofocus` behavior in Safari for `<sl-dialog>` and `<sl-drawer>` [#693](https://github.com/shoelace-style/shoelace/issues/693)
+- Improved type to select logic in `<sl-menu>` so it supports <kbd>Backspace</kbd> and gives users more time before resetting
 - Removed feature detection for `focus({ preventScroll })` since it no longer works in Safari
+- Removed the `--sl-tooltip-arrow-start-end-offset` design token
+- Replaced Popper positioning dependency with Floating UI in `<sl-dropdown>` and `<sl-tooltip>`
 
 ## 2.0.0-beta.70